requirements.txt
vllm
transformers
pandas
pyyaml
python-dotenv
nltk
rouge
tqdm